+++ chirho/dynamical/handlers/__init__.py
@@ -1,5 +1,5 @@
 from ..internals.solver import Solver  # noqa: F401
-from .event_loop import SimulatorEventLoop  # noqa: F401
+from .event_loop import InterruptionEventLoop  # noqa: F401
 from .interruption import (  # noqa: F401
     DynamicInterruption,
     DynamicIntervention,
