+++ chirho/dynamical/ops_/__init__.py
@@ -1,0 +1,1 @@
+from .ops import InPlaceDynamics, State, Trajectory, simulate  # noqa: F401
